requirements.txt
flask
flask_api
psycopg2-binary
configparser
requests
json